Rename widget_container.py to widget.py
