+++ mpfmc/widgets/bezier.py
@@ -4,7 +4,7 @@ from kivy.graphics.context_instructions import Color, Rotate, Scale
 from kivy.properties import (ListProperty, NumericProperty, OptionProperty,
                              BooleanProperty)
 
-from mpfmc.uix.widget_container import ContainedWidget
+from mpfmc.uix.widget import ContainedWidget
 from mpfmc.core.utils import center_of_points_list
 
 if TYPE_CHECKING:
