+++ mpfmc/widgets/text_input.py
@@ -11,7 +11,7 @@ from collections import deque
 from typing import TYPE_CHECKING, Optional
 from kivy.clock import Clock
 
-from mpfmc.uix.widget_container import WidgetContainer
+from mpfmc.uix.widget import WidgetContainer
 from mpfmc.widgets.text import Text
 
 if TYPE_CHECKING:
